+++ src/preprocessing/mod.rs
@@ -6,6 +6,7 @@
 
 pub mod auto_type;
 pub mod binarizer;
+pub mod constant_column_remover;
 pub mod encoder;
 pub mod feature_hasher;
 pub mod imputer;